boq_management_app/requirements.txt
streamlit
pandas
psycopg2-binary
openpyxl
python-dotenv
